==> 7_random_forest.ipynb <==
# Random Forest

import pandas as pd
import numpy as np
from tabulate import tabulate
from sklearn.ensemble import RandomForestClassifier
from sklearn.model_selection import RandomizedSearchCV
from sklearn.metrics import confusion_matrix, classification_report


# Load the data
train = pd.read_csv('data/train_rose.csv')
test = pd.read_csv('data/test.csv')

X_train = train.drop('CARAVAN', axis=1)
y_train = train['CARAVAN']
X_test = test.drop('CARAVAN', axis=1)
y_test = test['CARAVAN']

# Define a param grid for the randomForest

# Number of trees in random forest
n_estimators = [int(x) for x in np.linspace(start = 200, stop = 1000, num = 5)]
# Number of features to consider at every split
max_features = ['sqrt', 'log2']
# Maximum number of levels in tree
max_depth = [int(x) for x in np.linspace(10, 50, num = 5)]
# Minimum number of samples required to split a node
min_samples_split = [2, 5, 10]
# Minimum number of samples required at each leaf node
min_samples_leaf = [1, 2, 4]

# Create the random grid
random_grid = {'n_estimators': n_estimators,
               'max_features': max_features,
               'max_depth': max_depth,
               'min_samples_split': min_samples_split,
               'min_samples_leaf': min_samples_leaf
               }

cv = RandomizedSearchCV(estimator=RandomForestClassifier(), param_distributions=random_grid, n_iter=100, cv=3, verbose=10, random_state=42, n_jobs=-1)
cv.fit(X_train, y_train)

# Test the best model on the test set
best_model = cv.best_estimator_
y_pred = best_model.predict(X_test)

# Confusion matrix and classification report
# Tabulate the confusion matrix and write that columns are the predicted values and rows are the true values
conf = confusion_matrix(y_test, y_pred)
print('\n\nConfusion matrix\n')
print(tabulate(conf, headers=['Predicted 0', 'Predicted 1'], showindex=['True 0', 'True 1'], tablefmt='pretty'))

# Print the classification report
print('\n\nClassification report\n')
print(classification_report(y_test, y_pred))



#Print the best parameters
cv.best_params_

# Get the feature importances
feature_importances = best_model.feature_importances_
# Print the feature importances sorted by importance
print('\n\nFeature importances\n')
print(tabulate(sorted(zip(X_train.columns, feature_importances), key=lambda x: x[1], reverse=True), headers=['Feature', 'Importance'], tablefmt='pretty'))

# Fit a decision tree using only the first 5 most important features
from sklearn.tree import DecisionTreeClassifier
# Get the 5 most important features
top5_features = [x[0] for x in sorted(zip(X_train.columns, feature_importances), key=lambda x: x[1], reverse=True)[:10]]
# Fit the model
tree = DecisionTreeClassifier()
tree.fit(X_train[top5_features], y_train)
# Get the predictions
y_pred_tree = tree.predict(X_test[top5_features])
# Confusion matrix and classification report
# Tabulate the confusion matrix and write that columns are the predicted values and rows are the true values
conf_tree = confusion_matrix(y_test, y_pred_tree)
print('\n\nConfusion matrix for the decision tree\n')
print(tabulate(conf_tree, headers=['Predicted 0', 'Predicted 1'], showindex=['True 0', 'True 1'], tablefmt='pretty'))
print('\n\nClassification report for the decision tree\n')
print(classification_report(y_test, y_pred_tree))
