==> 7_random_forest.Rmd <==
# Random Forest

```{r}
library(randomForest)
library(caret)
```

Load the data and convert the CARAVAN column to a factor
```{r}
train <- read.csv("data/train_rose.csv")
test <- read.csv("data/test.csv")
train$CARAVAN <- as.factor(train$CARAVAN)
test$CARAVAN <- as.factor(test$CARAVAN)
```

```{r}
cv <- trainControl(method = "cv", number = 3)
```


```{r}
rf_model <- train(CARAVAN ~ .,
                  data = train,
                  method = "rf",
                  trControl = cv,
                  metric = "Accuracy",
                  tuneLength = 3,
                  verboseIter = TRUE)
```

```{r}
x_test <- test[, names(test) != "CARAVAN"]
pred <- predict(rf_model, x_test)
```

```{r}
confusionMatrix(pred, test$CARAVAN, positive = "1")
```



```{r}
print(rf_model)
```
